# openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/stores/{storeId}/orders/checkout": {
      "post": {
        "operationId": "OrdersController_checkout",
        "summary": "Checkout — create order from active cart",
        "parameters": [
          {
            "name": "storeId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Orders"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/stores/{storeId}/orders": {
      "get": {
        "operationId": "OrdersController_listOrders",
        "summary": "List orders",
        "parameters": [
          {
            "name": "storeId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Orders"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/stores/{storeId}/orders/{orderNumber}": {
      "get": {
        "operationId": "OrdersController_getOrder",
        "summary": "Get order by order number",
        "parameters": [
          {
            "name": "storeId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "orderNumber",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Orders"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/stores/{storeId}/orders/{id}/status": {
      "patch": {
        "operationId": "OrdersController_updateStatus",
        "summary": "Update order status",
        "parameters": [
          {
            "name": "storeId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Orders"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/stores/{storeId}/orders/{id}/fulfill": {
      "post": {
        "operationId": "OrdersController_createFulfillment",
        "summary": "Create fulfillment",
        "parameters": [
          {
            "name": "storeId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Orders"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/stores/{storeId}/orders/{id}/fulfillments/{fulfillmentId}": {
      "patch": {
        "operationId": "OrdersController_updateFulfillment",
        "summary": "Update fulfillment",
        "parameters": [
          {
            "name": "storeId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "fulfillmentId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Orders"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
    "/api/stores/{storeId}/payments/initiate": {
      "post": {
        "operationId": "PaymentsController_initiatePayment",
        "summary": "Initiate payment for an order",
        "parameters": [
          {
            "name": "storeId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Payment initiated"
          }
        },
        "tags": [
          "Payments"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/stores/{storeId}/payments/callback/kaspi": {
      "post": {
        "operationId": "PaymentsController_kaspiCallback",
        "summary": "Kaspi Pay callback endpoint",
        "parameters": [
          {
            "name": "storeId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Callback processed"
          }
        },
        "tags": [
          "Payments"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/stores/{storeId}/payments/callback/halyk": {
      "post": {
        "operationId": "PaymentsController_halykCallback",
        "summary": "Halyk Bank callback endpoint",
        "parameters": [
          {
            "name": "storeId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Callback processed"
          }
        },
        "tags": [
          "Payments"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/stores/{storeId}/payments/{id}/refund": {
      "post": {
        "operationId": "PaymentsController_refund",
        "summary": "Process refund (full or partial)",
        "parameters": [
          {
            "name": "storeId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Refund processed"
          }
        },
        "tags": [
          "Payments"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/stores/{storeId}/payments/simulate/{orderId}": {
      "post": {
        "operationId": "PaymentsController_simulatePayment",
        "summary": "Simulate successful payment (demo only)",
        "parameters": [
          {
            "name": "storeId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "orderId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Payment simulated and confirmed"
          }
        },
        "tags": [
          "Payments"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/stores/{storeId}/payments/order/{orderId}": {
      "get": {
        "operationId": "PaymentsController_getPaymentsByOrder",
        "summary": "Get all payments for an order",
        "parameters": [
          {
            "name": "storeId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "orderId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of payments"
          }
        },
        "tags": [
          "Payments"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
  },
  "info": {
    "title": "ShopBuilder API",
    "description": "ShopBuilder — Shopify Clone for Local Kazakh Merchants. E-commerce platform API.",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "access-token": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "description": "Paste JWT access token (without \"Bearer \" prefix)"
      }
    },
...
  }
}